% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.data.R
\name{generate.data}
\alias{generate.data}
\title{Generate simulated data}
\usage{
generate.data (p, n, umin, umax)
}
\arguments{
\item{p}{The number of genes.}

\item{n}{The sample size.}

\item{umin}{The lower limits of the edge values.}

\item{umax}{The upper limits of the edge values.}
}
\value{
\item{\code{X}}{ A matrix of sample  matrice  (\eqn{2n \times p}) from two different conditions.}
\item{\code{group}}{ A matrix of sample label matrice  (\eqn{2n \times 1}) from two different conditions.}
\item{\code{rho}}{A list (length = 2) of the partical coefficients matrices (\eqn{p \times p}).}
}
\description{
The complete procedure for generating simulated data. For details, refer to simulation study (Section 3.1 in the main text).
}
\details{
The function is used to generate the gene expression datasets.
}
\examples{
# Simulation data
data.x = generate.data(p = 100, n = 100, umin = 0.5, umax = 1)
}
\references{
Jia-Juan Tu, Le Ou-Yang, Hong Yan, Hong Qin and Xiao-Fei Zhang(2020), Differential network analysis
 by simultaneously considering  changes in  gene interactions and gene expression.
}
\seealso{
{ \code{\link{chNet}}, \code{\link{TCGA.BRCA}}, \code{\link{GSE13159.AML}}}
}
\author{
Jia-Juan Tu
}
