% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataAML.R
\docType{data}
\name{GSE13159.AML}
\alias{GSE13159.AML}
\title{Acute myeloid leukemia data (GSE13159)}
\format{An object of class \code{list} of length 2.}
\source{
[1] The Cancer Genome Atlas Research Network (2012), Comprehensive molecular portraits of human breast tumors. Nature. 490 (7418), 61-70. (http://cancergenome.nih.gov/).
}
\usage{
data("GSE13159.AML")
}
\description{
Gene expression datasets for acute myeloid leukemia (GSE13159) is download from http://discern-leelab.cs.washington.ed.
There are 541 AML and 73 normal samples.
The data include expression measurements of genes that overlap
with the mTOR signaling pathway (hsa04150)  from the Kyoto Encyclopedia of
Genes and Genomes databas and 50 non-differentially expressed  genes
(Student's t-test, p-value > 0.05) with the highest variation AML samples.
It includes an expression matrix for which the rows represent the
samples and the columns represent the genes, and a vector denoting the group membership of each
sample.
}
\examples{
data("GSE13159.AML")
}
\references{
Jia-Juan Tu, Le Ou-Yang, Hong Yan, Hong Qin and Xiao-Fei Zhang (2021), Differential network analysis
 by simultaneously considering  changes in  gene interactions and gene expression.
}
\seealso{
{ \code{\link{generate.data}},  \code{\link{chNet}},\code{\link{TCGA.BRCA}}}
}
\author{
Jia-Juan Tu
}
\keyword{datasets}
