% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chNet.R
\name{chNet}
\alias{chNet}
\title{Differential network analysis by simultaneously considering  changes in gene interactions and gene expression}
\usage{
chNet (X,group,lambar,parallel,nCpus)
}
\arguments{
\item{X}{Data matrix for which the rows represent the samples and the columns represent the genes.}

\item{group}{Vector which defines two groups of samples under comparison.}

\item{lambar}{The tuning (threshold) parameter controls the sparsity level.}

\item{parallel}{Logical value to indicate if the process should be run parallelly in multiple threads, default to FALSE.}

\item{nCpus}{number of (maximum) cores to use for parallel execution, default to 4.}
}
\value{
\item{\code{diff.edge}}{the adjacency matrix of the estimated differential network.}
\item{\code{diff.gene}}{the adjacency matrix of the estimated differentially expressed genes.}
\item{\code{Diff.net}}{the estimated differential network over all genes.}
\item{\code{Diff.net.connected}}{the estimated differential network over only the connected genes.}
}
\description{
The complete procedure for estimating differential network
 using chNet. For details, refer to method (Section 2.4 in the main text).
}
\details{
This function is implemented to infer diffferential network that satisfy hierarchichal constraints.
We first define the differential network as the difference of partial correlations,
and develop a new  test statistic to quantify the change of partial correlations.
Then the Student's t-test statistic is used to quantify the changes in expression levels of individual genes.
Finally, an optimization model is developed to combine the two different types of test statistics so that the
estimated differential networks exhibit the hierarchical structures.
A closed-formed solution is derived to solve the optimization model.
}
\examples{
# Simulation data
data.x= generate.data(p = 100, n = 100, umin = 0.5, umax = 1)
result = chNet(data.x$X,data.x$group, lambar = 4, parallel = FALSE, nCpus = 4)

# TCGA breast cancer data
data("TCGA.BRCA")
result = chNet(TCGA.BRCA$X,TCGA.BRCA$group, lambar = 2.85, parallel = FALSE, nCpus = 4)
# GSE13159 AML
data("GSE13159.AML")
result = chNet(TCGA.BRCA$X,TCGA.BRCA$group, lambar = 2.7, parallel = FALSE, nCpus = 4)
}
\references{
Jia-Juan Tu, Le Ou-Yang, Hong Yan, Hong Qin and Xiao-Fei Zhang(2020), Differential network analysis
 by simultaneously considering  changes in  gene interactions and gene expression.
}
\seealso{
{ \code{\link{generate.data}}, \code{\link{TCGA.BRCA}}, \code{\link{GSE13159.AML}}}
}
\author{
Jia-Juan Tu
}
